vk_bot/requirements.txt
fastapi
httpx
uvicorn[standard]
python-dotenv
mypy
ruff
